add reset buttons, fix some bugs

+++ examples/add_grayscale_image.py
@@ -4,8 +4,9 @@ Display one image using the add_image API.
 
 from skimage import data
 import napari
+import numpy as np
 
 
 with napari.gui_qt():
     # create the viewer with an image
-    viewer = napari.view_image(data.astronaut().mean(-1).astype('int'))
+    viewer = napari.view_image(data.astronaut().mean(-1).astype(np.uint16))
